[core] Move hoistInternalStatics into core
