+++ packages/material-ui/src/utils/hoistInternalStatics.js
